tests/wpt/tests/css/css-conditional/container-queries/style-change-in-container.html

<!doctype html>
<title>CSS Container Queries Test: recompute style inside a @container</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #container { container-type: size; }
  @container (min-width: 1px) {
    #content { color: green; }
  }
</style>
<div id="container">
  <div id="content"></div>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  let content = document.getElementById("content");

  test(() => {
    assert_equals(getComputedStyle(content).color, "rgb(0, 128, 0)");

    // Dirty style of an element inside the container:
    content.style.backgroundColor = "lime";

    // The container query should still evaluate correctly:
    assert_equals(getComputedStyle(content).color, 'rgb(0, 128, 0)');
  }, "Basic test for container query evaluation stability");
</script>
