LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/style-change-in-container-expected.txt


FAIL Basic test for container query evaluation stability assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
